Separated played_sound into it's own file
