+++ c_version/src/timer.c
@@ -6,17 +6,6 @@
 #include <time.h>
 #include "../include/timer.h"
 
-#define HOME_FOLDER_UNIX "~"
-#define HOME_FOLDER_WINDOWS "\%userprofile\%"
-#define UNIX_FILENAME ".local/assets/duck_quack.mp3"
-#define WIN_FILENAME ".local\\assets\\duck_quack.mp3"
-#define MAC_PLAYER "afplay"
-#define WIN_PLAYER "mpg123"
-#define LINUX_PLAYER "mpg123"
-#define MAX_COMMAND_LENGTH 100
-#define UNIX_NULL_OUTPUT " >/dev/null 2>&1"
-#define WIN_NULL_OUTPUT " >NUL 2>&1"
-
 void checkforEndSwitch(char* argument, bool* quietMode, bool* dryRun);
 int convertHoursMinsToSeconds(int hours, int minutes, int seconds);
 int convertArgsToSeconds(char* hoursString, char* minutesString, char* secondsString);
@@ -25,7 +14,6 @@ int checkArgument(char* input);
 int promptTimer();
 int convertToSeconds(char* input);
 void runTimer (int seconds, bool dryRunMode);
-void alert(int times, bool quietMode);
 void convert12to24(char* timeInput);
 void convert24ClockTo12(char* clockTime);
 
@@ -330,46 +318,5 @@ void convert24ClockTo12(char* clockTime){
     }
 
     strcat(clockTime, timerSuffix);
-}
-
-// Not Tested
-void alert(int times, bool quietMode){
-    int i;
-    char soundCommand[MAX_COMMAND_LENGTH];
-    char homeFolder[MAX_COMMAND_LENGTH];
-    char unixSeparator = '/';
-    char winSeparator = '\\';
-
-    printf("\rTime's up!\n");
-    if (quietMode == true)
-        exit(0);
-
-    for (i = 0; i < times; i++){
-        #if defined(__APPLE__)
-            strcpy (soundCommand, MAC_PLAYER);
-            strcat (soundCommand," ");
-            strcat (soundCommand, HOME_FOLDER_UNIX);
-            strncat (soundCommand, &unixSeparator, 1);
-            strcat (soundCommand,UNIX_FILENAME);
-            strcat (soundCommand,UNIX_NULL_OUTPUT);
-            system(soundCommand);
-        #elif defined(__linux__)
-            strcpy (soundCommand, LINUX_PLAYER);
-            strcat (soundCommand," ");
-            strcat (soundCommand, HOME_FOLDER_UNIX);
-            strncat (soundCommand, &unixSeparator, 1);
-            strcat (soundCommand,UNIX_FILENAME);
-            strcat (soundCommand,UNIX_NULL_OUTPUT);
-            system(soundCommand);
-        #elif defined(_WIN32)
-            strcpy (soundCommand, WIN_PLAYER);
-            strcat (soundCommand," ");
-            strcat (soundCommand, HOME_FOLDER_WINDOWS);
-            strncat (soundCommand, &winSeparator, 1);
-            strcat (soundCommand,WIN_FILENAME);
-            strcat (soundCommand,WIN_NULL_OUTPUT);
-            system(soundCommand);
-        #endif
-    }
 }
